+++ xroad-catalog-collector/src/main/java/fi/vrk/xroad/catalog/collector/actors/FetchCompaniesActor.java
@@ -14,7 +14,6 @@ package fi.vrk.xroad.catalog.collector.actors;
 
 import fi.vrk.xroad.catalog.collector.util.OrganizationUtil;
 import fi.vrk.xroad.catalog.collector.wsimport.ClientType;
-import fi.vrk.xroad.catalog.collector.wsimport.XRoadClientIdentifierType;
 import fi.vrk.xroad.catalog.persistence.CatalogService;
 import fi.vrk.xroad.catalog.persistence.CompanyService;
 import fi.vrk.xroad.catalog.persistence.entity.*;
@@ -36,6 +35,9 @@ public class FetchCompaniesActor extends XRoadCatalogActor {
     @Value("${xroad-catalog.fetch-companies-url}")
     private String fetchCompaniesUrl;
 
+    @Value("${xroad-catalog.fetch-companies-limit}")
+    private Integer fetchCompaniesLimit;
+
     @Autowired
     protected CatalogService catalogService;
 
@@ -51,12 +53,24 @@ public class FetchCompaniesActor extends XRoadCatalogActor {
     protected boolean handleMessage(Object message) {
         if (message instanceof ClientType) {
             ClientType clientType = (ClientType) message;
-            XRoadClientIdentifierType client = clientType.getId();
-            log.info("Fetching data for company with businessCode {}", client.getMemberCode());
-            String businessCode = clientType.getId().getMemberCode();
-            JSONObject companyJson = OrganizationUtil.getCompany(clientType, fetchCompaniesUrl, businessCode, catalogService);
-            saveData(companyJson.optJSONArray("results"));
-            log.info("Successfully saved data for company with businessCode {}", businessCode);
+            JSONObject companiesJson = OrganizationUtil.getCompanies(clientType, fetchCompaniesLimit, fetchCompaniesUrl, catalogService);
+            JSONArray companiesArray = companiesJson.optJSONArray("results");
+            int numberOfCompanies = companiesArray.length();
+            log.info("Fetching data for {} companies", numberOfCompanies);
+            companiesArray.forEach(item -> {
+                JSONObject company = (JSONObject) item;
+                String businessCode = company.optString("businessId"); 
+                JSONObject companyJson = OrganizationUtil.getCompany(clientType, fetchCompaniesUrl, businessCode, catalogService);
+                saveData(companyJson.optJSONArray("results"));
+            });
+            
+            // for (int i = 0; i < numberOfCompanies; i++) {
+            //     JSONObject company = companiesArray.optJSONObject(i);
+            //     String businessCode = company.optString("businessId");  
+            //     JSONObject companyJson = OrganizationUtil.getCompany(clientType, fetchCompaniesUrl, businessCode, catalogService);
+            //     saveData(companyJson.optJSONArray("results"));
+            // }
+            log.info("Successfully saved data for {} companies", numberOfCompanies);
             return true;
         } else {
             return false;