+++ xroad-catalog-collector/src/main/java/fi/vrk/xroad/catalog/collector/configuration/ProductionConfiguration.java
@@ -12,13 +12,11 @@
  */
 package fi.vrk.xroad.catalog.collector.configuration;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 
 @Configuration
 @Profile({"production", "sshtest"} )
-@Slf4j
 public class ProductionConfiguration extends ApplicationConfiguration {
 
     // dont override anything from ApplicationConfiguration